feat(wsl): add support for WSL
Might change host name, `blitzar`, to something else.

Since the theme of my config is `Space` (originally thought about `Workspace`), the host names are
based on this list of `Cosmic objects` or `Space objects`,
https://www.popularmechanics.com/space/deep-space/g22130573/space-objects-ranked-by-how-cool-they-sound/.

The benefit of the `Space` theme is that the wallpapers can easily be adjusted to be more oriented
towards a certain color.

However the theme and host names may be changed in the future if I find a more fitting concept.
Like from this list of `Mythological places`: https://en.wikipedia.org/wiki/List_of_mythological_places

Signed-off-by: Fredrik Dahlström <fredahl71@gmail.com>