feat: macOS v15.1 broke kitty terminal, attempt at fixing issue
In the macOS v15.1 update the kitty terminal emulator broke and could not be opened with Raycast
(app launcher). Kitty can still be opened via Finder and via other terminals as a command. Couldn't
resolve the issue with overlays but I'll keep it for now keep the overlay in case for future
reference. Modifying inputs, versions, sources, etc. did not help. In this process, setup WezTerm,
this includes theme, font, etc. Still have to resolve not being able to type symbols like: {}[]\|
within WezTerm.

Signed-off-by: Fredrik Dahlström <fredahl71@gmail.com>