fix(shell): resolve major WakaTime-shell issue
WakaTime had a zsh issue where if a value/file is not found it causes an "exit" which should not
happen. The CLI tool for SurrealDB (surreal) also caused issued with homebrew and is such also fixed
(in this case by removal). Neovim overlay modified as tree-sitter-bundled updated. Lastly,
flake.lock updated.

Signed-off-by: Fredrik Dahlström <fredahl71@gmail.com>